SFEperl-mime-types: bump to 2.13
--HG--
branch : spec_SFEperl-mime-types

+++ specs/SFEperl-mime-types.spec
@@ -4,7 +4,8 @@
 %define build510 %( if [ -x /usr/perl5/5.10/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build512 %( if [ -x /usr/perl5/5.12/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build516 %( if [ -x /usr/perl5/5.16/bin/perl ]; then echo '1'; else echo '0'; fi)
-%define build520 %( if [ -x /usr/perl5/5.20/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define build522 %( if [ -x /usr/perl5/5.22/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define enable_test %( if [ "x${PERL_DISABLE_TEST}" = 'xtrue' ]; then echo '0'; else echo '1'; fi )
 %define include_executable 0
 
 %define cpan_name MIME-Types
@@ -14,8 +15,8 @@
 Summary:               Definition of MIME types
 Name:                  SFEperl-%{sfe_cpan_name}
 IPS_package_name:      library/perl-5/%{ips_cpan_name}
-Version:               2.12
-IPS_component_version: 2.12
+Version:               2.13
+IPS_component_version: 2.13
 License:               perl_5
 URL:                   https://metacpan.org/pod/MIME::Types
 Source0:               http://cpan.metacpan.org/authors/id/M/MA/MARKOV/MIME-Types-%{version}.tar.gz
@@ -30,9 +31,15 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-584
 Summary:          Definition of MIME types
 BuildRequires:    runtime/perl-584 = *
 BuildRequires:    library/perl-5/extutils-makemaker-584
+%if %{enable_test}
+BuildRequires:    library/perl-5/pathtools-584
+BuildRequires:    library/perl-5/scalar-list-utils-584
+BuildRequires:    library/perl-5/test-simple-584
+%endif
 Requires:         runtime/perl-584 = *
-Requires:         library/perl-5/list-util-584
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/pathtools-584
+Requires:         library/perl-5/scalar-list-utils-584
 Requires:         library/perl-5/test-simple-584
 
 %description 584
@@ -45,9 +52,13 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-510
 Summary:          Definition of MIME types
 BuildRequires:    runtime/perl-510 = *
 BuildRequires:    library/perl-5/extutils-makemaker-510
+BuildRequires:    library/perl-5/pathtools-510
+BuildRequires:    library/perl-5/scalar-list-utils-510
+BuildRequires:    library/perl-5/test-simple-510
 Requires:         runtime/perl-510 = *
-Requires:         library/perl-5/list-util-510
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/pathtools-510
+Requires:         library/perl-5/scalar-list-utils-510
 Requires:         library/perl-5/test-simple-510
 
 %description 510
@@ -60,9 +71,15 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-512
 Summary:          Definition of MIME types
 BuildRequires:    runtime/perl-512 = *
 BuildRequires:    library/perl-5/extutils-makemaker-512
+%if %{enable_test}
+BuildRequires:    library/perl-5/pathtools-512
+BuildRequires:    library/perl-5/scalar-list-utils-512
+BuildRequires:    library/perl-5/test-simple-512
+%endif
 Requires:         runtime/perl-512 = *
-Requires:         library/perl-5/list-util-512
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/pathtools-512
+Requires:         library/perl-5/scalar-list-utils-512
 Requires:         library/perl-5/test-simple-512
 
 %description 512
@@ -75,34 +92,47 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-516
 Summary:          Definition of MIME types
 BuildRequires:    runtime/perl-516 = *
 BuildRequires:    library/perl-5/extutils-makemaker-516
+Requires:         library/perl-5/%{ips_cpan_name}
+%if %{enable_test}
+BuildRequires:    library/perl-5/pathtools-516
+BuildRequires:    library/perl-5/scalar-list-utils-516
+BuildRequires:    library/perl-5/test-simple-516
+%endif
 Requires:         runtime/perl-516 = *
-Requires:         library/perl-5/list-util-516
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/pathtools-516
+Requires:         library/perl-5/scalar-list-utils-516
 Requires:         library/perl-5/test-simple-516
 
 %description 516
 Definition of MIME types
 %endif
 
-%if %{build520}
-%package 520
-IPS_package_name: library/perl-5/%{ips_cpan_name}-520
+%if %{build522}
+%package 522
+IPS_package_name: library/perl-5/%{ips_cpan_name}-522
 Summary:          Definition of MIME types
-BuildRequires:    runtime/perl-520 = *
-BuildRequires:    library/perl-5/extutils-makemaker-520
-Requires:         runtime/perl-520 = *
-Requires:         library/perl-5/list-util-520
-Requires:         library/perl-5/pathtools-520
-Requires:         library/perl-5/test-simple-520
+BuildRequires:    runtime/perl-522 = *
+BuildRequires:    library/perl-5/extutils-makemaker-522
+%if %{enable_test}
+BuildRequires:    library/perl-5/pathtools-522
+BuildRequires:    library/perl-5/scalar-list-utils-522
+BuildRequires:    library/perl-5/test-simple-522
+%endif
+Requires:         runtime/perl-522 = *
+Requires:         library/perl-5/%{ips_cpan_name}
+Requires:         library/perl-5/pathtools-522
+Requires:         library/perl-5/scalar-list-utils-522
+Requires:         library/perl-5/test-simple-522
 
-%description 520
+%description 522
 Definition of MIME types
 %endif
 
 
 %prep
 %setup -q -n %{cpan_name}-%{version}
-rm -rf %{buildroot}
+[ -d %{buildroot} ] && rm -rf %{buildroot}
 
 %build
 build_with_makefile.pl_for() {
@@ -115,8 +145,12 @@ build_with_makefile.pl_for() {
     ${bindir}/perl Makefile.PL PREFIX=%{_prefix} \
                    DESTDIR=$RPM_BUILD_ROOT \
                    LIB=${vendor_dir}
-    make
-    [ ${test} = 'without_test' ] || make test
+
+    export CC='cc -m32'
+    export LD='cc -m32'
+    echo ${perl_ver} | egrep '5\.(84|12)' > /dev/null || (export CC='cc -m64'; export LD='cc -m64')
+    make CC="${CC}" LD="${LD}"
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC="${CC}" "LD=${LD}"
     make pure_install
 }
 
@@ -131,17 +165,47 @@ build_with_build.pl_for() {
                    --installdirs vendor \
                    --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build
-    [ ${test} = 'without_test' ] || ${bindir}/perl ./Build test
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || ${bindir}/perl ./Build test
     ${bindir}/perl ./Build install --destdir $RPM_BUILD_ROOT
+    ${bindir}/perl ./Build clean
 }
 
 modify_bin_dir() {
-  perl_ver=$1
-  if [ -d $RPM_BUILD_ROOT/usr/bin ]
-  then
-    [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver} ] || mkdir -p $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
-    mv $RPM_BUILD_ROOT/usr/bin $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin
-  fi
+    perl_ver=$1
+    if [ -d $RPM_BUILD_ROOT/usr/bin ]
+    then
+      [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver} ] || mkdir -p $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+      mv $RPM_BUILD_ROOT/usr/bin $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin
+    fi
+      
+    if [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin ]
+    then
+        for i in $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin/*
+        do
+            sed -i.bak -e "s!/usr/bin/env perl!/usr/perl5/${perl-ver}/bin/perl!" ${i}
+            [ -f ${i}.bak] || rm -f ${i}.bak
+        done
+    fi
+}
+
+modify_man_dir() {
+    perl_ver=$1
+    if [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man ]
+    then
+        if [ -d $RPM_BUILD_ROOT%{_datadir}/man ]
+        then
+            rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
+        else
+            mkdir -p $RPM_BUILD_ROOT%{_datadir}
+            mv $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man $RPM_BUILD_ROOT%{_datadir}/
+            rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
+        fi
+        if [ %{include_executable} -eq 0 ]
+        then
+            rmdir $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+        fi
+
+    fi
 }
 
 build_for() {
@@ -153,7 +217,8 @@ build_for() {
     build_with_makefile.pl_for $*
   fi
 
-    modify_bin_dir $*
+  modify_bin_dir $*
+  modify_man_dir $*
 }
 
 # To build without test, pass 'without_test' to build_for commaond.
@@ -174,14 +239,14 @@ build_for 5.12
 build_for 5.16
 %endif
 
-%if %{build520}
-build_for 5.20
+%if %{build522}
+build_for 5.22
 %endif
 
 %install
-mkdir -p $RPM_BUILD_ROOT%{_datadir}
 if [ -d $RPM_BUILD_ROOT%{_prefix}/man ]
 then
+    mkdir -p $RPM_BUILD_ROOT%{_datadir}
     mv $RPM_BUILD_ROOT%{_prefix}/man $RPM_BUILD_ROOT%{_datadir}
 fi
 if [ -d $RPM_BUILD_ROOT%{_datadir}/man/man3 ]
@@ -236,18 +301,19 @@ rm -rf %{buildroot}
 %endif
 %endif
 
-%if %{build520}
-%files 520
+%if %{build522}
+%files 522
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
-/usr/perl5/vendor_perl/5.20
+/usr/perl5/vendor_perl/5.22
 %if %{include_executable}
-/usr/perl5/5.20
+/usr/perl5/5.22
 %endif
 %endif
-
 
 %changelog
+* Mon Jun 05 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- bump to 2.13
 * Thu Nov 12 2015 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - bump to 2.12 and build packages for perl-510, perl-516 and perl-520
 * Fri Jun 22 2012 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
